Several improvements
Now store the parameters as a list. The job_id is passed to each worker
to retrieve the correct arguments. This should allow a job to be stopped
and restarted by condor.
Several improvements to the Jobs class.
More efficient sorting of the jobs in the map() function
Don't allow to map lambda functions.